Fix paths and remove compiler packages
