+++ src/BeFaster/packages.config
@@ -2,8 +2,6 @@
 <packages>
   <package id="Apache.NMS" version="1.7.1" targetFramework="net45" />
   <package id="Apache.NMS.ActiveMQ" version="1.7.2" targetFramework="net45" />
-  <package id="Microsoft.CodeDom.Providers.DotNetCompilerPlatform" version="1.0.7" targetFramework="net45" />
-  <package id="Microsoft.Net.Compilers" version="2.3.2" targetFramework="net45" developmentDependency="true" />
   <package id="Newtonsoft.Json" version="9.0.1" targetFramework="net45" />
   <package id="RestSharp" version="105.2.3" targetFramework="net45" />
   <package id="TDL.Client" version="0.14.6" targetFramework="net45" />
